install.sh: by default, do not init git repository

+++ install.sh
@@ -27,6 +27,8 @@ for opt in "$@"; do
             no_zsh=true;;
         '--no-tmux')
             no_tmux=true;;
+        '--init-git')
+            init_git=true;;
         *)
             printf "%s\n" "Available options:" "--no-fonts" "--no-apt-setup" \
                 "--no-nvim" "--no-packages" "--no-symlinks" "--no-zsh" \
@@ -206,17 +208,19 @@ install_powerline_fonts() {
 
 init_git() {
     # Link downloaded dotfiles directory to the git repository
-    cd "$DOTFILES_DIR"
-    if ! is_git_repository; then
-        e_header "Initializing git repository..."
-        git init
-        git remote add origin $DOTFILES_GIT_REMOTE
-        git fetch origin master
-        git reset --hard FETCH_HEAD
-        git clean -fd # Remove any untracked files
-        git push -u origin master || e_error "Couldn't run 'git push -u origin master'"
+    if [[ $init_git ]]; then
+        cd "$DOTFILES_DIR"
+        if ! is_git_repository; then
+            e_header "Initializing git repository..."
+            git init
+            git remote add origin $DOTFILES_GIT_REMOTE
+            git fetch origin master
+            git reset --hard FETCH_HEAD
+            git clean -fd # Remove any untracked files
+            git push -u origin master || e_error "Couldn't run 'git push -u origin master'"
+        fi
+        cd -
     fi
-    cd -
 }
 
 setup_zsh() {
